+++ lib/index.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-import{b as o,c as u,e as p}from"./chunk.QIOKNTKE.js";import l from"esbuild";import c from"readline";let e;const r=c.createInterface({input:process.stdin,output:process.stdout}),n=async()=>{const i=o();return i.clear&&console.clear(),e?.kill(),new Promise(t=>{i.run?(r.pause(),e=p(i),e.on("close",()=>{r.resume(),i.watch?t():process.exit()})):(console.log(`${i.entry||i._.join(", ")} built successfully on ${i.outdir}`),r.close(),t())})},d=o(),a=async()=>{const i=await u(),t=d.watch&&await l.startService(),s=t?t.build:l.build;return async()=>await Promise.all([s(i),n()])},m=async()=>{const i=o();if(i.version){const{displayVersion:s}=await import("./displayVersion.js");return s()}const t=await a();i.watch?(await import("chokidar")).watch(i.watch).on("ready",async()=>await t()).on("change",async()=>await t()):await t()};m();
+import{b as o,c as u,e as p}from"./chunk.ZAHFTB6G.js";import l from"esbuild";import c from"readline";let e;const r=c.createInterface({input:process.stdin,output:process.stdout}),n=async()=>{const i=o();return i.clear&&console.clear(),e?.kill(),new Promise(t=>{i.run?(r.pause(),e=p(i),e.on("close",()=>{r.resume(),i.watch?t():process.exit()})):(console.log(`${i.entry||i._.join(", ")} built successfully on ${i.outdir}`),r.close(),t())})},d=o(),a=async()=>{const i=await u(),t=d.watch&&await l.startService(),s=t?t.build:l.build;return async()=>await Promise.all([s(i),n()])},m=async()=>{const i=o();if(i.version){const{displayVersion:s}=await import("./displayVersion.js");return s()}const t=await a();i.watch?(await import("chokidar")).watch(i.watch).on("ready",async()=>await t()).on("change",async()=>await t()):await t()};m();
