+++ src/initEsbuild.ts
@@ -1,8 +1,7 @@
 import esbuild from 'esbuild'
 import type { BuildResult } from 'esbuild'
 import { postBuild } from 'postBuild'
-import { getEsbuildConfig } from 'helpers'
-import getCLIOptions from 'helpers/getOptions'
+import { getCLIOptions, getEsbuildConfig } from 'helpers'
 
 // Simplify this type
 type InitEsbuild = () => Promise<() => Promise<[BuildResult, void]>>
