+++ src/validateEntryPoints.ts
@@ -1,6 +1,5 @@
 import isGlob from 'is-glob'
-import { globParser } from 'helpers'
-import { fileExists } from 'helpers'
+import { globParser, fileExists } from 'helpers'
 import type { CLIFlags } from 'types'
 
 type ValidateEntryPoints = (options: Readonly<CLIFlags>) => Promise<string[]>
